refactor: split kernels into spheroidal and gridding submodules
