+++ idg/python/src/idg_python/gridder.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .kernels import (
+from .kernels.gridding import (
     add_subgrid_to_grid,
     compute_phasor,
     visibilities_to_subgrids,
