+++ idg/python/src/idg_python/taper.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .kernels import evaluate_spheroidal
+from .kernels.spheroidal import evaluate_spheroidal
 
 
 def get_taper(subgrid_size: int) -> np.ndarray:
